Module tclsci --> dynamic library
add setSCIpath and getSCIpath C functions to known SCI path
getSCIpath replaces getenv("SCI") in C

add get_cf_type and set_cf_type to know cf_type in a another (dll) module.

+++ scilab/modules/tclsci/sci_gateway/gw_tclsci.c
@@ -69,7 +69,7 @@ int C2F(gw_tclsci)()
 #ifndef _DEBUG
 			_try
 			{
-				(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+				(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 			}
 			_except (EXCEPTION_EXECUTE_HANDLER)
 			{	
@@ -78,10 +78,10 @@ int C2F(gw_tclsci)()
 				if (ExceptionString) {FREE(ExceptionString);ExceptionString=NULL;}
 			}
 #else
-			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 #endif
 #else
-			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,strlen(Tab[Fin-1].name));
+			(*(Tab[Fin-1].f)) (Tab[Fin-1].name,(unsigned long)strlen(Tab[Fin-1].name));
 #endif
 		}
 		else
